Return entire cart for all endpoints
